% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thorn.R
\name{thorn-shiny}
\alias{thorn-shiny}
\alias{thornOutput}
\alias{renderThorn}
\title{Shiny bindings for \code{thorn}}
\usage{
thornOutput(outputId, width = "100\%", height = "100\%")

renderThorn(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{a valid CSS measurement (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended}

\item{expr}{an expression that generates a shader created with
\code{\link{thorn}}}

\item{env}{the environment in which to evaluate \code{expr}}

\item{quoted}{logical, whether \code{expr} is a quoted expression}
}
\description{
Output and render functions for using \code{\link{thorn}} within Shiny
applications and interactive Rmd documents.
}
